backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "pdfminer-six (>=20250506,<20250507)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.82.0,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
django = "^5.2.1"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
